yeah

+++ src/main/java/com/techelevator/CashBox.java
@@ -1,17 +1,17 @@
 package com.techelevator;
 
 public class CashBox {
-    private int balance;
+    private Double balance;
 
 
     public String returnChange() {
         //converts balance into change and returns balance to zero
-        int quarters =  (balance * 100) / 25;
+        int quarters = (int) (balance * 100) / 25;
         balance = (balance * 100) % 25;
-        int dimes = balance /10;
+        int dimes = (int) (balance /10);
         balance = balance % 100;
-        int nickels = balance / 5;
-        balance = 0;
+        int nickels = (int) (balance / 5);
+        balance = 0.0;
         return (quarters + " quarters," + nickels + " nickels," + dimes + " dimes,");
     }
 
@@ -19,11 +19,11 @@ public class CashBox {
 
 
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
 }
